Disable seed script's save function temporarily
Change script so it doesn't save Mongoose documents to database for now.

Previously, for-loop saved document to DB per iteration. However,
JavaScript heap ran out of memory if the loop tried to save 200,000
records.

The for-loop has been altered to push those documents into an array.
The loop can create and push 500,000 to 1 million records into the array.